Rename config_loader to a more generic json_loader
