+++ article_generator/ai_chat.py
@@ -11,7 +11,7 @@ from anthropic import (
     APIConnectionError as ClaudeAPIConnectionError,
     APIError as ClaudeAPIError,
 )
-from utils.config_loader import load_config
+from utils.json_loader import load_config
 from logging_setup import setup_logger
 
 log = setup_logger(__name__)
